refactor(management): centralize request validation
